Rename ping test to reply_to_ping test
